+++ pkg/cortex/serve/cortex_internal/lib/api/__init__.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cortex_internal.lib.api.realtime import RealtimeAPI, get_api
-from cortex_internal.lib.api.utils import get_spec
+from cortex_internal.lib.api.realtime import RealtimeAPI
+from cortex_internal.lib.api.batch import BatchAPI
 from cortex_internal.lib.api.task import TaskAPI
+from cortex_internal.lib.api.utils import get_spec, model_downloader, DynamicBatcher, CortexMetrics
