+++ pkg/cortex/serve/cortex_internal/lib/api/batch.py
@@ -12,130 +12,40 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import itertools
-import threading as td
-import time
-import traceback
-from collections import defaultdict
-from http import HTTPStatus
-from typing import Any, Callable, Dict, List
+from copy import deepcopy
+import imp
+import inspect
+import os
+import dill
+from typing import Any, Dict, Optional, Union
 
-from starlette.responses import Response
+from datadog.dogstatsd.base import DogStatsd
 
-from cortex_internal.lib.exceptions import UserRuntimeException
-from cortex_internal.lib.log import logger
-
-
-class DynamicBatcher:
-    def __init__(
-        self,
-        predictor_impl: Callable,
-        max_batch_size: int,
-        batch_interval: int,
-        test_mode: bool = False,
-    ):
-        self.predictor_impl = predictor_impl
-
-        self.batch_max_size = max_batch_size
-        self.batch_interval = batch_interval  # measured in seconds
-        self.test_mode = test_mode  # only for unit testing
-        self._test_batch_lengths = []  # only when unit testing
-
-        self.barrier = td.Barrier(self.batch_max_size + 1)
-
-        self.samples = {}
-        self.predictions = {}
-        td.Thread(target=self._batch_engine, daemon=True).start()
-
-        self.sample_id_generator = itertools.count()
-
-    def _batch_engine(self):
-        while True:
-            if len(self.predictions) > 0:
-                time.sleep(0.001)
-                continue
-
-            try:
-                self.barrier.wait(self.batch_interval)
-            except td.BrokenBarrierError:
-                pass
-
-            self.predictions = {}
-            sample_ids = self._get_sample_ids(self.batch_max_size)
-            try:
-                if self.samples:
-                    batch = self._make_batch(sample_ids)
-
-                    predictions = self.predictor_impl.predict(**batch)
-                    if not isinstance(predictions, list):
-                        raise UserRuntimeException(
-                            f"please return a list when using server side batching, got {type(predictions)}"
-                        )
-
-                    if self.test_mode:
-                        self._test_batch_lengths.append(len(predictions))
-
-                    self.predictions = dict(zip(sample_ids, predictions))
-            except Exception as e:
-                self.predictions = {sample_id: e for sample_id in sample_ids}
-                logger.error(traceback.format_exc())
-            finally:
-                for sample_id in sample_ids:
-                    del self.samples[sample_id]
-                self.barrier.reset()
-
-    def _get_sample_ids(self, max_number: int) -> List[int]:
-        if len(self.samples) <= max_number:
-            return list(self.samples.keys())
-        return sorted(self.samples)[:max_number]
-
-    def _make_batch(self, sample_ids: List[int]) -> Dict[str, List[Any]]:
-        batched_samples = defaultdict(list)
-        for sample_id in sample_ids:
-            for key, sample in self.samples[sample_id].items():
-                batched_samples[key].append(sample)
-
-        return dict(batched_samples)
-
-    def _enqueue_request(self, sample_id: int, **kwargs):
-        """
-        Enqueue sample for batch inference. This is a blocking method.
-        """
-
-        self.samples[sample_id] = kwargs
-        try:
-            self.barrier.wait()
-        except td.BrokenBarrierError:
-            pass
-
-    def predict(self, **kwargs):
-        """
-        Queues a request to be batched with other incoming request, waits for the response
-        and returns the prediction result. This is a blocking method.
-        """
-        sample_id = next(self.sample_id_generator)
-        self._enqueue_request(sample_id, **kwargs)
-        prediction = self._get_prediction(sample_id)
-        return prediction
-
-    def _get_prediction(self, sample_id: int) -> Any:
-        """
-        Return the prediction. This is a blocking method.
-        """
-        while sample_id not in self.predictions:
-            time.sleep(0.001)
-
-        prediction = self.predictions[sample_id]
-        del self.predictions[sample_id]
-
-        if isinstance(prediction, Exception):
-            return Response(
-                content=str(prediction),
-                status_code=HTTPStatus.INTERNAL_SERVER_ERROR,
-                media_type="text/plain",
-            )
-
-        return prediction
+from cortex_internal.lib import util
+from cortex_internal.lib.client.python import PythonClient
+from cortex_internal.lib.client.tensorflow import TensorFlowClient
+from cortex_internal.lib.model import (
+    FileBasedModelsGC,
+    ModelTreeUpdater,
+    ModelsGC,
+    TFSAPIServingThreadUpdater,
+    ModelsTree,
+    ModelsHolder,
+)
+from cortex_internal.lib.api.utils import CortexMetrics, model_downloader
+from cortex_internal.lib.api.validations import (
+    are_models_specified,
+    validate_class_impl,
+    validate_predictor_with_grpc,
+    validate_python_predictor_with_models,
+)
+from cortex_internal.lib.type import (
+    PythonPredictorType,
+    TensorFlowNeuronPredictorType,
+    TensorFlowPredictorType,
+    predictor_type_from_api_spec,
+)
+from cortex_internal.lib.exceptions import CortexException, UserException, UserRuntimeException
 
 PYTHON_CLASS_VALIDATION = {
     "required": [
@@ -147,7 +57,7 @@ PYTHON_CLASS_VALIDATION = {
         {
             "name": "handle_batch",
             "required_args": ["self"],
-            "optional_args": ["payload", "batch_id"]
+            "optional_args": ["payload", "batch_id"],
         },
     ],
     "optional": [
@@ -155,10 +65,7 @@ PYTHON_CLASS_VALIDATION = {
             "name": "load_model",
             "required_args": ["self", "model_path"],
         },
-        {
-            "name": "on_job_complete",
-            "required_args": ["self"]
-        },
+        {"name": "on_job_complete", "required_args": ["self"]},
     ],
 }
 
@@ -172,30 +79,29 @@ TENSORFLOW_CLASS_VALIDATION = {
         {
             "name": "handle_batch",
             "required_args": ["self"],
-            "optional_args": ["payload", "batch_ud"]
+            "optional_args": ["payload", "batch_ud"],
         },
     ],
     "optional": [
-        {
-            "name": "on_job_complete",
-            "required_args": ["self"]
-        },
+        {"name": "on_job_complete", "required_args": ["self"]},
     ],
 }
 
 
-class BatchingAPI:
+class BatchAPI:
     """
     Class to validate/load the predictor class (PythonPredictor, TensorFlowPredictor).
     Also makes the specified models in cortex.yaml available to the predictor's implementation.
     """
 
-    def __init__(self, api_spec: dict, model_dir: str):
+    def __init__(self, api_spec: dict, statsd_client: DogStatsd, model_dir: str):
         """
         Args:
             api_spec: API configuration.
             model_dir: Where the models are stored on disk.
         """
+
+        self.metrics = CortexMetrics(statsd_client, api_spec)
 
         self.type = predictor_type_from_api_spec(api_spec)
         self.path = api_spec["predictor"]["path"]
@@ -436,109 +342,3 @@ class BatchingAPI:
             cron.stop()
         for cron in self.crons:
             cron.join()
-
-
-class RealtimeAPI:
-    def __init__(
-        self,
-        storage: S3,
-        api_spec: Dict[str, Any],
-        model_dir: str,
-        cache_dir: str = ".",
-    ):
-        self.storage = storage
-        self.api_spec = api_spec
-        self.cache_dir = cache_dir
-
-        self.id = api_spec["id"]
-        self.predictor_id = api_spec["predictor_id"]
-        self.deployment_id = api_spec["deployment_id"]
-
-        self.key = api_spec["key"]
-        self.metadata_root = api_spec["metadata_root"]
-        self.name = api_spec["name"]
-        self.predictor = Predictor(api_spec, model_dir)
-
-        host_ip = os.environ["HOST_IP"]
-        datadog.initialize(statsd_host=host_ip, statsd_port=9125)
-        self.__statsd = datadog.statsd
-
-    @property
-    def statsd(self):
-        return self.__statsd
-
-    @property
-    def python_server_side_batching_enabled(self):
-        return (
-            self.api_spec["predictor"].get("server_side_batching") is not None
-            and self.api_spec["predictor"]["type"] == "python"
-        )
-
-    def metric_dimensions_with_id(self):
-        return [
-            {"Name": "api_name", "Value": self.name},
-            {"Name": "api_id", "Value": self.id},
-            {"Name": "predictor_id", "Value": self.predictor_id},
-            {"Name": "deployment_id", "Value": self.deployment_id},
-        ]
-
-    def metric_dimensions(self):
-        return [{"Name": "api_name", "Value": self.name}]
-
-    def post_request_metrics(self, status_code, total_time):
-        total_time_ms = total_time * 1000
-        metrics = [
-            self.status_code_metric(self.metric_dimensions(), status_code),
-            self.status_code_metric(self.metric_dimensions_with_id(), status_code),
-            self.latency_metric(self.metric_dimensions(), total_time_ms),
-            self.latency_metric(self.metric_dimensions_with_id(), total_time_ms),
-        ]
-        self.post_metrics(metrics)
-
-    def post_status_code_request_metrics(self, status_code):
-        metrics = [
-            self.status_code_metric(self.metric_dimensions(), status_code),
-            self.status_code_metric(self.metric_dimensions_with_id(), status_code),
-        ]
-        self.post_metrics(metrics)
-
-    def post_latency_request_metrics(self, total_time):
-        total_time_ms = total_time * 1000
-        metrics = [
-            self.latency_metric(self.metric_dimensions(), total_time_ms),
-            self.latency_metric(self.metric_dimensions_with_id(), total_time_ms),
-        ]
-        self.post_metrics(metrics)
-
-    def post_metrics(self, metrics):
-        try:
-            if self.statsd is None:
-                raise CortexException("statsd client not initialized")  # unexpected
-
-            for metric in metrics:
-                tags = ["{}:{}".format(dim["Name"], dim["Value"]) for dim in metric["Dimensions"]]
-                if metric.get("Unit") == "Count":
-                    self.statsd.increment(metric["MetricName"], value=metric["Value"], tags=tags)
-                else:
-                    self.statsd.histogram(metric["MetricName"], value=metric["Value"], tags=tags)
-        except:
-            logger.warn("failure encountered while publishing metrics", exc_info=True)
-
-    def status_code_metric(self, dimensions, status_code):
-        status_code_series = int(status_code / 100)
-        status_code_dimensions = dimensions + [
-            {"Name": "response_code", "Value": "{}XX".format(status_code_series)}
-        ]
-        return {
-            "MetricName": "cortex_status_code",
-            "Dimensions": status_code_dimensions,
-            "Value": 1,
-            "Unit": "Count",
-        }
-
-    def latency_metric(self, dimensions, total_time):
-        return {
-            "MetricName": "cortex_latency",
-            "Dimensions": dimensions,
-            "Value": total_time,  # milliseconds
-        }
