+++ pkg/cortex/serve/start/task.py
@@ -12,15 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
 import os
 import sys
 from copy import deepcopy
 
 from cortex_internal.lib import util
 from cortex_internal.lib.api import get_spec, TaskAPI
-from cortex_internal.lib.exceptions import UserRuntimeException
-from cortex_internal.lib.telemetry import capture_exception, get_default_tags, init_sentry
+from cortex_internal.lib.telemetry import get_default_tags, init_sentry
 from cortex_internal.lib.log import configure_logger
 
 init_sentry(tags=get_default_tags())
